New package types created and files moved to it.
Package types has been created. All ic_types (both python and cython)
are now moved to this package.

In addition all modules related with event_modules are now in package evm.
