+++ invisible_cities/cities/dorothea.py
@@ -10,7 +10,7 @@ from .. core.configure         import configure
 from .. core.system_of_units_c import units
 
 from .. io.kdst_io              import kr_writer
-from .. reco.event_model        import PersistentKrEvent
+from .. evm.event_model         import PersistentKrEvent
 
 from .. reco                   import tbl_functions   as tbl
 from .. reco                   import pmaps_functions_c as pmp
