+++ invisible_cities/cities/irene_test.py
@@ -19,7 +19,7 @@ from pytest import fixture
 from .  irene import Irene
 from .. core                 import system_of_units as units
 from .. core.configure       import configure
-from .. core.ic_types        import minmax
+from .. types.ic_types       import minmax
 from .. io.pmap_io           import read_run_and_event_from_pmaps_file
 from .. reco.params          import S12Params as S12P
