+++ invisible_cities/io/kdst_io_test.py
@@ -9,7 +9,7 @@ from ..core.test_utils    import assert_dataframes_equal
 from ..reco.dst_functions import load_dst
 from . kdst_io            import kr_writer
 from . kdst_io            import xy_writer
-from ..reco.event_model   import PersistentKrEvent
+from ..evm.event_model   import PersistentKrEvent
 
 
 def test_Kr_writer(config_tmpdir, Kr_dst_data):
