+++ invisible_cities/io/pmap_io_test.py
@@ -24,7 +24,7 @@ from .. reco                   import peak_functions_c as cpf
 from .. reco.params            import S12Params        as S12P
 from .. reco.params            import ThresholdParams
 from .. reco.params            import PMaps
-from .. core.ic_types          import minmax
+from .. types.ic_types         import minmax
 
 from . pmap_io                 import pmap_writer
 from .. evm.pmaps              import S1
